move hook to hooks place
