+++ shared/hooks/index.js
@@ -2,3 +2,4 @@ import useMutationObserver from './useMutationObserver';
 
 export { useMutationObserver };
 export * from './useWindowTitle';
+export * from './useSideDrawer.js';
